Refactor: ショップ詳細ページのリファクタを実施
